requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.5.0
pyserial>=3.5
pyusb>=1.2
